fix(full-boot): wave-3 Hibernate schema alignment for clinical services
Map document OCR confidence to NUMERIC, inventory-elmis outbox payload to
jsonb, and pacs imaging_study metadata to jsonb while keeping TEXT columns
where Flyway defines them.

+++ services/document-service/src/main/java/zw/gov/mohcc/impilo/docstore/core/ocr/OcrService.java
@@ -10,6 +10,7 @@ import zw.gov.mohcc.impilo.shared.auth.TrustContext;
 import zw.gov.mohcc.impilo.shared.auth.TrustContextHolder;
 
 import java.io.InputStream;
+import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 import java.util.LinkedHashMap;
 import java.util.Map;
@@ -53,7 +54,7 @@ public class OcrService {
             OcrProvider.OcrResult result = provider.extract(objectId, content, object.getMimeType());
             job.setStatus("COMPLETED");
             job.setExtractedText(result.extractedText());
-            job.setConfidence(result.confidence());
+            job.setConfidence(BigDecimal.valueOf(result.confidence()));
             job.setCompletedAt(OffsetDateTime.now());
             return toPayload(ocrJobRepository.save(job));
         } catch (Exception ex) {
